requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.6.3
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cssselect==1.2.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
forex-python==1.8
fpdf==1.7.2
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.21.4
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
panda==0.3.1
pandas==2.1.3
parso==0.8.3
pathlib==1.0.1
pikepdf==8.9.0
Pillow==10.1.0
platformdirs==4.2.0
playsound==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyforest==1.1.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==4.6
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sgmllib3k==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.16.0
youtube-transcript-api==0.6.2
